Index, Kostenverlauf hinzugefügt
