+++ package.json
@@ -20,17 +20,9 @@
         "axios": "^0.19.2",
         "chart.js": "^2.9.3",
         "chartjs-plugin-annotation": "^0.5.7",
-        "chartjs-plugin-datalabels": "^0.7.0",
         "chartjs-plugin-zoom": "^0.7.5",
-        "d3": "^5.15.0",
-        "dateformat": "^3.0.3",
-        "html2canvas": "^1.0.0-rc.5",
-        "html2pdf.js": "^0.9.2",
-        "jspdf": "^1.5.3",
-        "moment": "^2.24.0",
         "quasar": "^1.9.9",
         "vue-chartjs": "^3.5.0",
-        "vue-i18n": "^8.16.0",
         "vue-lottie": "^0.2.1"
     },
     "devDependencies": {
