+++ src/components/Dashboard/HistoryChartStacked.vue
@@ -7,11 +7,11 @@
         borderless
         v-model="model"
         dense
-        :options="selectOption"
+        :options="selects"
         map-options
         emit-value
         label="Auswahl"
-        style="width: 100px"
+        style="width: 120px"
       ></q-select>
     </div>
     <commit-chart-bar :width="w" :height="h" :chartData="datacollection" :options="options"></commit-chart-bar>
@@ -21,6 +21,7 @@
 <script>
 import CommitChartBar from "../js/CommitChartBar.js";
 import { mapGetters } from "vuex";
+import { interpolateColors } from "../js/InterpolateColors";
 
 export default {
   name: "HistChart",
@@ -30,7 +31,7 @@ export default {
   data() {
     return {
       model: "parts",
-      selectOption: [
+      selects: [
         {
           label: "Material",
           value: "parts"
@@ -46,6 +47,14 @@ export default {
         {
           label: "Maschine",
           value: "machines"
+        },
+        {
+          label: "Produktgruppe",
+          value: "productgrp"
+        },
+        {
+          label: "Werkstoff",
+          value: "material"
         }
       ],
       datacollection: {
@@ -62,6 +71,21 @@ export default {
   },
   props: ["tab"],
   computed: {
+    Data() {
+      let data = [];
+      switch (this.tab) {
+        case "intern":
+          data = this.History;
+          break;
+        case "extern":
+          data = this.HistoryExtern;
+          break;
+        case "all":
+          data = this.HistoryAll;
+          break;
+      }
+      return data;
+    },
     ...mapGetters({
       History: "dataset/getHistory",
       HistoryAll: "dataset/getHistoryAll",
@@ -70,7 +94,9 @@ export default {
   },
   watch: {
     model() {
-      this.fillData();
+      if (typeof this.Data === "object" && Object.keys(this.Data).length > 0) {
+        this.fillData();
+      }
     }
   },
   methods: {
@@ -87,18 +113,8 @@ export default {
       const cm = "rgb(217, 221, 3)";
       const c2 = "rgb(221, 3, 51)";
 
-      let dataHistory = [];
-      switch (this.tab) {
-        case "intern":
-          dataHistory = JSON.parse(this.History[this.model]);
-          break;
-        case "extern":
-          dataHistory = JSON.parse(this.HistoryExtern[this.model]);
-          break;
-        case "all":
-          dataHistory = JSON.parse(this.HistoryAll[this.model]);
-          break;
-      }
+      const dataHistory = JSON.parse(this.Data[this.model]);
+
       if (Object.keys(dataHistory).length === 0) {
         return;
       }
@@ -108,9 +124,9 @@ export default {
       const s1 = Math.floor(steps / 2);
       const s2 = steps - s1;
 
-      colorOne = this.interpolateColors(c1, cm, s1);
+      colorOne = interpolateColors(c1, cm, s1);
 
-      colorTwo = this.interpolateColors(cm, c2, s2);
+      colorTwo = interpolateColors(cm, c2, s2);
 
       colors = colorOne.concat(colorTwo);
 
@@ -199,35 +215,6 @@ export default {
           }
         }
       };
-    },
-    // Funktionen interpolateColor und interpolateColors für die Farben im gestapelten Balkendiagramm
-    // https://github.com/ondras/rot.js
-    interpolateColor(color1, color2, factor) {
-      if (arguments.length < 3) {
-        factor = 0.5;
-      }
-      const result = color1.slice();
-      for (let i = 0; i < 3; i++) {
-        result[i] = Math.round(result[i] + factor * (color2[i] - color1[i]));
-      }
-      return result;
-    },
-
-    interpolateColors(color1, color2, steps) {
-      const stepFactor = 1 / (steps - 1);
-      const interpolatedColorArray = [];
-
-      // \d => einfache Stelle (56 => 5,6) \d+ => zwei Stellen (56=>56)
-      // Number Konstruktor stellt Konvertierung in eine Zahl sicher
-      color1 = color1.match(/\d+/g).map(Number);
-      color2 = color2.match(/\d+/g).map(Number);
-
-      for (let i = 0; i < steps; i++) {
-        interpolatedColorArray.push(
-          this.interpolateColor(color1, color2, stepFactor * i)
-        );
-      }
-      return interpolatedColorArray;
     }
   },
   mounted() {
