+++ src/components/MenuItem.vue
@@ -6,9 +6,7 @@
 
     <q-item-section>
       <q-item-label>{{ title }}</q-item-label>
-      <q-item-label caption>
-        {{ caption }}
-      </q-item-label>
+      <q-item-label caption>{{ caption }}</q-item-label>
     </q-item-section>
   </q-item>
 </template>
@@ -41,9 +39,15 @@ export default {
 </script>
 
 <style lang="sass">
+.my-menu-link
+  background-color: $info
 .menu-link
   color: white
   background: #F2C037
 .mini-icon
   color: $secondary
+.q-item:hover
+  transform: scale(1.1)
+.q-item
+  transition: 0.1s
 </style>
